+++ firebase-vscode/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## NEXT
 
+- [Fixed] Data Connect emulator issues properly streamed on startup
+- [Fixed] Data Connect schema reloads consistently
+
 ## 0.4.0
 
 - Updated internal firebase-tools dependency to 13.13.0